% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcity_function.R
\name{get_city}
\alias{get_city}
\title{Get City Function}
\usage{
get_city(string)
}
\arguments{
\item{string}{The search string to find the city}
}
\description{
This function allows you to search by city name and returns the latitude and
longitude formatted for use in further API requests
}
\examples{
get_city()
}
\keyword{city}
\keyword{search}
