% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearby_search.R
\name{nearby_search}
\alias{nearby_search}
\title{Get Nearby Locations Function}
\usage{
nearby_search(location_name, category)
}
\arguments{
\item{location_name}{The search string to find the city}

\item{category}{The type of location to retrieve}
}
\value{
A tidy dataframe with 10 locations of chosen category and their location_id, name, distance and bearing
nearby_search()
}
\description{
This function allows you to search by place name and category of attraction (e.g. restaurant) and
returns 10 nearby attractions of that category and their distance from the location
}
