% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_ten.R
\name{top_ten}
\alias{top_ten}
\title{Top Ten Function}
\usage{
top_ten(string, type = NA, city = NA)
}
\arguments{
\item{string}{The search string for the query}

\item{type}{Optional field to limit the business type to one of "hotels",
"attractions", "restaurants", or "geos".}

\item{city}{Optional field that will limit search to a specific geo location}
}
\value{
10 obs. of 4 variables: id, name, country, address
}
\description{
This function returns IDs and names on the top ten best matches of businesses
based on a search string and optional business type and location filters
}
