% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_location_info.R
\name{get_location_info}
\alias{get_location_info}
\title{Get Location Info Function}
\arguments{
\item{location_id}{The location_id to search}
}
\value{
A tidy dataframe with information for the inputted location_id
}
\description{
This function allows the user to input a location ID
which will then provide information for the specified location
Users can optionally set the currency for results (default CAD)
}
\examples{
get_location_info()
}
